push up changes for now - refactoring tests module since its gotten quite large
